% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{bh}
\alias{bh}
\title{Benjamini–Hochberg procedure and  Benjamini–Yekutieli procedure}

\description{
Benjamini–Hochberg procedure for controlling the false discovery rate and Benjamini–Yekutieli procedure for controlling the false discovery rate under dependency.
}

\usage{
bh(Y, X, alpha)
}

\arguments{
\item{Y}{A \eqn{n\times q} Response matrix.}

\item{X}{A \eqn{n\times p} Numeric design matrix for the model.}

\item{alpha}{The significant level. Default is \code{alpha = 0.05}.}
}

\value{
A list
\itemize{
\item bestset - The selected subscript position of the response best subset.
}
}

\references{
    Benjamini, Y. and Hochberg, Y. (1995). Controlling the false discovery rate a practical and powerfulapproach to multiple testing. Journal of the Royal Statistical Society Series B, 57, 289–300.

    Benjamini, Y. and Yekutieli, D. (2001). The control of the false discovery rate in multiple testing under dependency. The Annals of Statistics, 29, 1165–1188.
}

\examples{
library("stats")
library("MASS")
library("Matrix")
n   <- 200
p   <- 5
q   <- 10
q0  <- 5
beta <- matrix(runif(p*q0),p,q0)
eps <- matrix(rnorm(n*q),n,q)
x <- matrix(rnorm(n*p),n,p)
y <- cbind(x\%*\%beta, matrix(0,n,q-q0)) + eps
fit <- bh(y,x,alpha=0.05)
fit$bestset

data(simulatedData)
y = simulatedData$Y
w = simulatedData$W
fit <- bh(y,w,alpha=0.05)
fit$bestset
}
