% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonf.R
\name{bonf}
\alias{bonf}
\title{Bonferroni test}

\description{
Bonferroni correction test.
}

\usage{
bonf(Y, X, alpha)
}

\arguments{
\item{Y}{A \eqn{n\times q} Response matrix.}

\item{X}{A \eqn{n\times p} Numeric design matrix for the model.}

\item{alpha}{The significant level. Default is \code{alpha = 0.05}.}
}

\value{
A list
\itemize{
\item decision.set - The estimation of the \eqn{\delta}.
\item bestset - The selected subscript position of the response best subset.
}
}

\references{
    Liu, X., Hu, J., and Liu, X. (2024). Random Projection-Based Response Best-Subset Selector for Ultrahigh Dimensional Multi-Outcome Data. Manuscript.
}

\examples{
library("stats")
library("MASS")
library("Matrix")
n   <- 200
p   <- 5
q   <- 10
q0  <- 5
beta <- matrix(runif(p*q0),p,q0)
eps <- matrix(rnorm(n*q),n,q)
x <- matrix(rnorm(n*p),n,p)
y <- cbind(x\%*\%beta, matrix(0,n,q-q0)) + eps
fit <- bonf(y,x,alpha=0.05)
fit$bestdecision
fit$bestset

data(simulatedData)
y = simulatedData$Y
w = simulatedData$W
fit <- bonf(y,w,alpha=0.05)
}
