% Generated by roxygen2: do not edit by hand
% Please edit documentation in data/
\name{datas}
\docType{data}
\title{Simulated dataset and a real dataset.}

\description{
    Simulated data from a high-dimensional multivariate linear model and a real breast cancer data.
}

\usage{
    data(simulatedData)
}

\examples{
    library("stats")
    library("MASS")
    library("Matrix")
    data(simulatedData)
    y = simulatedData$Y
    w = simulatedData$W
    fit <- rbsrp(y,w)

    data(breastdata)
    chrom = breastdata$chrom
}
