% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbsrp.R
\name{rbsrp}
\alias{rbsrp}
\title{RBS based on the random projection approach}

\description{
Select the response variables and estimate regression coefficients simultaneously for multivariate linear regression without covariance of responses based on the random projection approach.
}

\usage{
rbsrp(Y, X, steplen = 0.00125)
}

\arguments{
\item{Y}{A \eqn{n\times q} Response matrix.}

\item{X}{A \eqn{n\times p} Numeric design matrix for the model.}

\item{steplen}{The step length of \eqn{\alpha}. Default is \code{steplen=0.00125}.}
}

\value{
A list
\itemize{
\item bestset - The selected subscript position of the response best subset.
\item bestdecision - The estimation of the \eqn{\delta}.
}
}

\references{
    Liu, X., Hu, J., and Liu, X. (2024). Random Projection-Based Response Best-Subset Selector for Ultrahigh Dimensional Multi-Outcome Data. Manuscript.
}

\examples{
library("stats")
library("MASS")
library("Matrix")
n   <- 200
p   <- 5
q   <- 10
q0  <- 5
beta <- matrix(runif(p*q0),p,q0)
eps <- matrix(rnorm(n*q),n,q)
x <- matrix(rnorm(n*p),n,p)
y <- cbind(x\%*\%beta, matrix(0,n,q-q0)) + eps
fit <- rbsrp(y,x)
fit$bestdecision
fit$bestset

data(simulatedData)
y = simulatedData$Y
w = simulatedData$W
fit <- rbsrp(y,w)
}
